Increase time limit not timeout
Action Scheduler 2.0 reintroduced the loop for processing scheduled
actions, and constrained it to a time limit, instead of just processing
a batch of actions regardless of how long that batch took. As a result,
increase the batch size won't necessarily mean that many actions gets
processed, and we also need to increase the time limit.

We also don't need to increase the timeout period anymore, because even
the increase time limit is below the default timeout period.